Write constituent records to their own file
